render, audio: make the pixel and audio arithmetic architecture-independent
The CI matrix put this port in front of an arm64 toolchain for the first time
(macos-latest), so I went looking for arithmetic that could answer differently
there before trusting a golden digest from it. Two functions could.

The Go spec allows an implementation to evaluate a float `a*b + c` as one fused
instruction, keeping the full product and rounding once where the separate
operations round twice. There is no flag to turn that off -- it is permitted, not
a bug -- and amd64 codegen declines it while arm64 takes it. Grepping the
assembly rather than the source is the only way to see it:

    GOOS=darwin GOARCH=arm64 go build -gcflags=-S ./... 2>&1 | grep FMADD

That found three FMADDDs across two call sites.

FillPolyPatOrGray's scanline crossing is the serious one. The crossing is
rounded up to a pixel column, so a last-bit difference on a value sitting
exactly on a boundary does not shift a pixel by a hair -- it paints a whole
column that the other machine does not, because ceil(5.0) is 5 and
ceil(5.000000000000001) is 6. Four callers in objectdraw.go draw furniture
shadows into the background of nearly every room in the game, so an affected
polygon would differ in every screenshot of that room for the rest of the run.

Rather than patch the rounding I removed the floats. The vertices are integers
and the row centre y+0.5 doubles to an odd integer, so the crossing is a
rational with a small numerator and denominator: (2y+1-2ay)(bH-aH) over
2(by-ay), which with the ceil folded in is an integer division. ceilHalf(float64)
int becomes ceilDiv(n, den int64) int64, the row's whole inner loop is ints, and
there is no rounding left for two architectures to disagree about. int64 for the
product because an off-screen vertex can be five figures.

audio.stepFor keeps its float64 and gains a conversion around the product, which
forces the intermediate rounding and forbids the fusion. That is the cheap fix
and the right one here: the value is computed once per sound at load time, not
per pixel in an inner loop. Left unfixed, a header rate sitting on a boundary
would resample by one part in 65536 on an Apple Silicon Mac and not on an x86
one, and every mixed sample after it would differ.

No golden changed, and I checked that rather than assuming it. I patched both
sites to call math.FMA explicitly -- emulating what arm64 emits -- and ran the
pixel and audio suites (fidelity, render, game, shell, audio, replay). Every
digest matched. A separate 90-million-edge sweep over integer vertices confirms
the two forms genuinely diverge, 1,235,147 crossings landing in a different
column, so the hazard is real; it is simply that no polygon any shipped house
draws and no rate any shipped sound carries sits on one of those boundaries.

So this was latent, and I fixed it anyway. A suite that hashes pixels and audio
bytes has to be architecture-independent by construction: if it is only
accidentally so, the next vertex a new house introduces turns a latent
difference into a wrong screenshot that reproduces on one machine and not
another, and 4.5's bug report would faithfully record two digests with nothing
to explain them. docs/IMPROVEMENTS.md 4.9 has the whole finding, including the
standing FMADD check I have left as a note rather than written.

Two unrelated things ride along, both from the same CI run:

- Every workflow action is bumped past the Node 20 runtime GitHub has
  deprecated: checkout@v5, setup-go@v6, upload-artifact@v5,
  download-artifact@v5. Sixteen references across both files; the tags cannot be
  verified from this host, only read from the deprecation notice.

- .gitignore now covers /glidergo and /glidertool. `go build ./...` -- which the
  CI runs and which is the first thing anyone with the toolchain will type --
  drops one executable per main package into the working directory rather than
  into bin/, so the tree came back dirty with two untracked 14 MB files.

This does not explain the `go test ./...` failure on the native matrix. It was
found while looking for it, and it is not it.

+++ internal/audio/bank.go
@@ -185,11 +185,19 @@ const FixedOne = 1 << 16
 //
 // A rate of zero -- a manifest with no rate column, or a Sound built by hand in a test -- is the
 // Macintosh rate, because that is what every sound the game shipped with is.
+//
+// The conversion around the product is deliberate and must stay. Written `x*FixedOne + 0.5` the
+// compiler may contract the multiply and the add into one FMA -- the Go spec allows it, arm64
+// codegen takes it -- and an FMA rounds once where the two operations round twice. That is a
+// last-bit difference in a value this function then truncates, so a rate sitting on a boundary
+// would step by one 65536th more on an Apple Silicon Mac than on an x86 one, and every mixed
+// sample after it would differ. An explicit float64 conversion forces the intermediate rounding
+// and forbids the fusion. See docs/IMPROVEMENTS.md 4.9.
 func stepFor(rateHz float64) int64 {
 	if rateHz <= 0 {
 		return FixedOne
 	}
-	return int64(rateHz*RateDen/RateNum*FixedOne + 0.5)
+	return int64(float64(rateHz*RateDen/RateNum*FixedOne) + 0.5)
 }
 
 // Frames is the sample's length. It is the byte length, the sound being 8-bit mono, and it
